move most room calls off of the UI thread with livedata

+++ app/src/main/java/com/example/jgjio_desktop/gostats/ActiveListViewModel.java
@@ -24,7 +24,7 @@ public class ActiveListViewModel extends AndroidViewModel{
     //todo fix this cast from long to int
 
     //return the id
-    int insertStatisticalList(StatisticalList newList) { return (int) mRepository.insertStatisticalList(newList);}
+    long insertStatisticalList(StatisticalList newList) { return mRepository.insertStatisticalList(newList); }
 
     void deleteList(StatisticalList statList) {mRepository.removeStatisticalList(statList); }
